gitlab-telegram-notify-bot/requirements.txt
aiogram==3.7.0
requests==2.32.3
python-dotenv==1.0.1